+++ kt-string-similarity/src/commonMain/kotlin/ca/solostudios/stringsimilarity/Levenshtein.kt
@@ -3,7 +3,7 @@
  * Copyright (c) 2015-2015 Thibault Debatty
  *
  * The file Levenshtein.kt is part of kt-fuzzy
- * Last modified on 22-10-2021 08:01 p.m.
+ * Last modified on 22-10-2021 08:10 p.m.
  *
  * MIT License
  *
@@ -97,7 +97,7 @@ public class Levenshtein : MetricStringDistance {
      */
     public fun distance(s1: String, s2: String, limit: Int = Int.MAX_VALUE): Double {
         if (s1 == s2) {
-            return 0.toDouble()
+            return 0.0
         }
         if (s1.isEmpty()) {
             return s2.length.toDouble()
